NativeImage class...

+++ include/client/resources/metadata/animation/AnimationMetadataSectionSerializer.h
@@ -1,11 +1,11 @@
 #pragma once
 
-#include "client/resources/data/AnimationMetadataSection.h"
+#include "client/resources/metadata/animation/AnimationMetadataSection.h"
 #include "resources/data/IMetadataSectionSerializer.h"
 #include "util/JsonUtils.h"
 
 class AnimationMetadataSectionSerializer : IMetadataSectionSerializer<AnimationMetadataSection> {
-		AnimationFrame* parseAnimationFrame(u8 frame, cJSON* element);
+		AnimationFrame* parseAnimationFrame(u16 frame, cJSON* element);
 
 	public:
 		AnimationMetadataSection deserialize(cJSON* json) override;
