+++ src/client/resources/metadata/animation/AnimationMetadataSection.cpp
@@ -1,4 +1,4 @@
-#include "client/resources/data/AnimationMetadataSection.h"
+#include "client/resources/metadata/animation/AnimationMetadataSection.h"
 
 #include "client/Exception.h"
 #include "util/math/NumberUtils.h"
@@ -7,8 +7,8 @@ static bool isMultipleOf(int valMul, int val) {
 	return valMul / val * val == valMul;
 }
 
-Pair<u8, u8> AnimationMetadataSection::getSpriteSize(u8 width, u8 height) {
-	Pair<u8, u8> pair = getFrameSize(width, height);
+Pair<u16, u16> AnimationMetadataSection::getSpriteSize(u16 width, u16 height) {
+	Pair<u16, u16> pair = getFrameSize(width, height);
 
 	if (!isMultipleOf(width, pair.first) || !isMultipleOf(height, pair.second))
 		Crash("Image size %u, %u is not multiple of frame size %u, %u", width, height, pair.first, pair.second);
@@ -16,17 +16,17 @@ Pair<u8, u8> AnimationMetadataSection::getSpriteSize(u8 width, u8 height) {
 	return pair;
 }
 
-Pair<u8, u8> AnimationMetadataSection::getFrameSize(u8 defWidth, u8 defHeight) {
+Pair<u16, u16> AnimationMetadataSection::getFrameSize(u16 defWidth, u16 defHeight) {
 	if (mFrameWidth != -1)
-		return mFrameHeight != -1 ? Pair<u8, u8>(mFrameWidth, mFrameHeight) : Pair<u8, u8>(mFrameWidth, defHeight);
+		return mFrameHeight != -1 ? Pair<u16, u16>(mFrameWidth, mFrameHeight) : Pair<u16, u16>(mFrameWidth, defHeight);
 	else if (mFrameHeight != -1)
-		return Pair<u8, u8>(defWidth, mFrameHeight);
+		return Pair<u16, u16>(defWidth, mFrameHeight);
 
 	int i = MIN(defWidth, defHeight);
-	return Pair<u8, u8>(static_cast<u8>(i), static_cast<u8>(i));
+	return Pair<u16, u16>(static_cast<u16>(i), static_cast<u16>(i));
 }
 
-u8 AnimationMetadataSection::getFrameTimeSingle(u8 frame) const {
+u16 AnimationMetadataSection::getFrameTimeSingle(u16 frame) const {
 	AnimationFrame animationFrame = getAnimationFrame(frame);
 	return animationFrame.hasNoTime() ? mFrameTime : animationFrame.getFrameTime();
 }
